巨大更新： 1. 更改为彩色图 2. 加入识别功能 3. 加入分颜色标记识别结果功能 待完成： 去粘连

+++ LearnCpp/connect.h
@@ -68,7 +68,7 @@ namespace im {
 		int relabel_image(Tlabel *out, int width, int height);
 
 		std::vector<Similarity> labels;
-		int highest_label;
+		unsigned int highest_label;
 	};
 
 	int ConnectedComponents::connected(const Tin *img, Tlabel *labelimg, int width, int height, bool K8_connectivity)
@@ -144,7 +144,7 @@ namespace im {
 	int ConnectedComponents::relabel_image(Tlabel *labelimg, int width, int height)
 	{
 		int newtag = 0;
-		for (int id = 0; id < labels.size(); ++id)
+		for (size_t id = 0; id < labels.size(); ++id)
 			if (is_root_label(id))
 				labels[id].tag = newtag++;
